Refactor CheckDeprecationHeader function and remove unused imports
