+++ httpclient/httpclient_helpers.go
@@ -2,28 +2,12 @@
 package httpclient
 
 import (
-	"net/http"
 	"time"
-
-	"github.com/deploymenttheory/go-api-http-client/logger"
-	"go.uber.org/zap"
 )
 
 // ParseISO8601Date attempts to parse a string date in ISO 8601 format.
 func ParseISO8601Date(dateStr string) (time.Time, error) {
 	return time.Parse(time.RFC3339, dateStr)
-}
-
-// CheckDeprecationHeader checks the response headers for the Deprecation header and logs a warning if present.
-func CheckDeprecationHeader(resp *http.Response, log logger.Logger) {
-	deprecationHeader := resp.Header.Get("Deprecation")
-	if deprecationHeader != "" {
-
-		log.Warn("API endpoint is deprecated",
-			zap.String("Date", deprecationHeader),
-			zap.String("Endpoint", resp.Request.URL.String()),
-		)
-	}
 }
 
 // RedactSensitiveData redacts sensitive data if the HideSensitiveData flag is set to true.
